fix(dispatcher,supply): re-root non-top auth entries; count unhandled meta versions; stop silently dropping unattributable removals
Cold audit 2026-08-04 of internal/dispatcher + the five supply observers.

AUTH-TREE RE-ROOTING (dispatcher). extractInvokeContractCallTrees
re-rooted auth subtrees under the top-level call only when NO walked
call matched it. In the mixed shape — a co-signed tx where entry 0
authorizes the top-level call and entry 1 authorizes a deeper call
signed by a different party — containsCall found entry 0's match, the
re-root pass was skipped, and entries 1..n kept CallPath nil, i.e. were
exported as the op's ENTRY POINT. Downstream that is call_depth 0 /
call_kind 'top_level', so a router wrapped by an aggregator is recorded
as the entry point and TagTradesRoutedVia's sub_invocation join misses,
under-reporting the real wrapper on /v1/aggregators. It also let two
calls in one op share the (TxHash, OpIndex, CallPath) identity that
ContractCallContext documents as a stable dedup key. Each entry is now
classified independently, with non-top entries re-rooted past the
top-level call's own sub-invocation space so paths cannot collide.
TestExtractCallTrees_multipleAuthEntries builds this exact XDR but
asserts only counts and function names, so the mislabelling was
invisible; the new test asserts the paths and is proven red against the
old logic.

META-VERSION DEFAULT ARM (dispatcher + the ClickHouse twin). Both walks
switched on TransactionMeta.V with arms for 3 and 4 and no default, so
an unhandled version skipped the whole apply phase silently. Now
counted (Stats.EntryMetaUnsupported / LedgerExtract.EntryMetaUnsupported,
WARNed by statsflush and the live sink) — the same treatment the sibling
tx-event path got in G15-06, and for the same reason: an unwalked apply
phase is indistinguishable from a ledger in which nothing happened.

This is defence, not a live gap, and two comments claiming otherwise are
corrected. The old note said V1/V2 "don't carry post-Soroban operation
metadata so their apply phase is skipped … which is correct". That
reason is false — both metas carry full per-operation
LedgerEntryChanges, and the SDK's own LedgerChangeReader handles them.
The real reason we never see them is that galexie's captive core
re-generates meta at replay time. Verified by decoding 48 production
ledgers from both datastores this code reads, protocols 1 to 19 and
8,000+ transactions: every tx carried V4, including ledger 1,000,023 at
protocol 1. r1's lake agrees — 70.5 billion apply-phase
ledger_entry_changes rows below the protocol-20 activation, back to
ledger 3. Also softened "mirrors the SDK exactly" to "follows", noting
the one real divergence (we omit the txInternalError suppression at
LedgerVersion <= 12, so we emit more rather than fewer).

UNATTRIBUTABLE REMOVALS (claimable_balances, liquidity_pools). Decode's
!known branch returned (nil, nil). Matches consults hasPreImage, which
does NOT apply lookupPreImage's ledger guard, so that branch is the
stale-memo case — and a dropped removal is not a dropped increment: the
removal is the only thing that stops the claimed balance being counted,
so the stale amount stays in the served total forever, over-reporting
supply with nothing moving. Now returns a wrapped sentinel so it lands
on the per-source decode-error counter. Both memo tests still assert
zero observations and additionally assert the drop is signalled.

int16 NARROWING. InsertSEP41SupplyEventBatch bound op_index/event_index
as int16 while the single-row writer binds int. An op emitting >32,767
supply events would wrap negative, violate CHECK (op_index >= 0) and
fail the ENTIRE batch rather than one row.

MIGRATION COMMENTS. 0013 listed XLM as a possible lp_reserve_observations
asset_key — the observer returns ErrUnsupportedLPAsset for native, so
native LP reserves are not covered by that table at all. 0014 now records
why G-strkey holders appear only for pure SEP-41 tokens, which is what
makes the Algorithm-2 decomposition non-overlapping.

+++ internal/storage/timescale/sep41_supply_events.go
@@ -566,8 +566,15 @@ func (s *Store) InsertSEP41SupplyEventBatch(ctx context.Context, rows []SEP41Sup
 			e.ContractID,
 			int64(e.Ledger),
 			e.TxHash,
-			int16(e.OpIndex),
-			int16(e.EventIndex),
+			// int, not int16: the columns are `integer`, and the
+			// single-row InsertSEP41SupplyEvent binds int. Narrowing here
+			// meant an op emitting >32,767 supply events wrapped negative
+			// and violated the `CHECK (op_index >= 0)` constraint, failing
+			// the ENTIRE batch rather than one row — an asymmetry between
+			// two writers of the same table that no lockstep test covers
+			// (cold audit 2026-08-04).
+			int(e.OpIndex),
+			int(e.EventIndex),
 			e.ObservedAt.UTC(),
 			string(e.Kind),
 			e.Amount.String(),
